Move loadconfigurations to a dedicated go file
